Diagnose.sh
Small script that fixes an issue where the clipboard cannot be detected in Linux Mint when using the Pyperclip library. Kind of important to have that working when a lot of the utility of some of these programs is the ability to directly copy things to your clipboard for automation. Unsure if this fixes clipboard behavior in VMs for now.