Calculator Improvements
The calculator app is now intended to "emulate" the TI84 in that, one should be able to use it fully in place of a TI84 or as if it was a TI84 on steroids (TI84 cannot do binomial theorem whereas dconsole calculator can). Thus the interface has been modified a little bit to mimic the TI84 in some ways. FOr example, there is a 2nd "button" - when you enter "2nd" in the command line, the options table presented to you at the top of the interface changes to show alternate options, highlighted in blue, for example, sin changes to arcsin when in the 2nd interface mode for the options table.

Standouts:
- New "range" command in calculator - intended to emulate the TI84's Y= button which then allows the user to insert functions and generate a table when they hit 2nd+graph (table). In dconsole, you can enter up to 4 functions and generate a TI84 like table of their values based on the x-value.
- Cleaned up some code in calculator, the line stop=input("Hit ENTER to continue ") now is at the end of the logic function, meaning this line does not appear at the end of every code block under an elif statement
- Mode command in calculator which allows you to choose between radians and degrees for trigonometric functions (more later)
- improved download.sh script
- dtools now has a function to reverse a string and another to reverse a list