Friendship ended with Pyperclip now Pyclip is my best friend
Got rid of Pyperclip for Pyclip because Pyperclip would randomly just not work which is hella annoying. Idk if Pyclip works on mobile yet because I still need to modify the code so that it can run in Termux.

All files have been set to lowercase naming, which makes it easier to run programs in dconsole etc.

Small UI changes in markdownwriter (still trying to find good names for some of these programs)